final_pipe/antiscam_model/predictor/src/requirements.txt
fastapi
uvicorn==0.17.6
joblib==1.3.2
numpy==1.23.5
scikit-learn==1.3.0
google-cloud-storage>=1.26.0,<2.0.0dev
google-cloud-aiplatform[prediction]>=1.16.0